Explicit conversions of doubles to ints in attack-pred and attack-prey.

+++ source/cpu/cHardwareExperimental.cc
@@ -3865,10 +3865,10 @@ bool cHardwareExperimental::Inst_AttackPrey(cAvidaContext& ctx)
   const int success_reg = FindModifiedRegister(rBX);   
   const int bonus_reg = FindModifiedNextRegister(success_reg);
   setInternalValue(success_reg, 1, true);   
-  setInternalValue(bonus_reg, (int) target_bonus, true);
+  setInternalValue(bonus_reg, (int) (target_bonus), true);
   const int spec_bin = m_organism->GetRBins()[m_world->GetConfig().COLLECT_SPECIFIC_RESOURCE.Get()];
   setInternalValue(FindModifiedNextRegister(bonus_reg), spec_bin, true);
-  setInternalValue(FindModifiedNextRegister(FindModifiedNextRegister(bonus_reg)), m_organism->GetRBinsTotal(), true);
+  setInternalValue(FindModifiedNextRegister(FindModifiedNextRegister(bonus_reg)), (int) (m_organism->GetRBinsTotal()), true);
   return true;
 } 		
 
@@ -4108,10 +4108,10 @@ bool cHardwareExperimental::Inst_AttackPred(cAvidaContext& ctx)
   const int success_reg = FindModifiedRegister(rBX);   
   const int bonus_reg = FindModifiedNextRegister(success_reg);
   setInternalValue(success_reg, 1, true);   
-  setInternalValue(bonus_reg, (int) target_bonus, true);
+  setInternalValue(bonus_reg, (int) (target_bonus), true);
   const int spec_bin = m_organism->GetRBins()[m_world->GetConfig().COLLECT_SPECIFIC_RESOURCE.Get()];
   setInternalValue(FindModifiedNextRegister(bonus_reg), spec_bin, true);
-  setInternalValue(FindModifiedNextRegister(FindModifiedNextRegister(bonus_reg)), m_organism->GetRBinsTotal(), true);
+  setInternalValue(FindModifiedNextRegister(FindModifiedNextRegister(bonus_reg)), (int) (m_organism->GetRBinsTotal()), true);
   return true;
 }
